% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score_array_DEV.R
\name{calculate_score_array_DEV}
\alias{calculate_score_array_DEV}
\title{calculate_score_array_DEV}
\usage{
calculate_score_array_DEV(
  MSspectrum = NULL,
  mz = NULL,
  intensity = NULL,
  sequence = NULL,
  PTMformula = "C0",
  charge = 1,
  range12C = c(0.986, 1),
  range14N = c(0.994, 1),
  range16O = c(0.996, 1),
  abund17O = c(0.00038),
  range32S = c(0.944, 0.953),
  abund33SCoarse = c(0.00749),
  range33S = c(0, 0.01),
  abundStepCoarse = 0.001,
  abundStepFine = 1e-04,
  SNR = 25,
  method = "MAD",
  refineMz = "kNeighbors",
  k = 2,
  resolvingPower = 3e+05,
  compFunc = "dotproduct",
  binSize = 0.05
)
}
\arguments{
\item{MSspectrum}{An MSnbase Spectrum1 object.}

\item{mz}{A numeric vector containing m/z values for a spectrum.}

\item{intensity}{A numeric vector containing intensity values for a spectrum.}

\item{sequence}{Sequence of the protein represented by the m/z and intensity vectors.}

\item{PTMformula}{Chemical formula of PTMs of the proteoform represented by the m/z and intensity vectors. Formulas for all PTMs should be combined.}

\item{charge}{Charge state of the proteoform represented by the m/z and intensity vectors.}

\item{abundStepCoarse}{Abundance step to use for the coarse score array Must be larger than abundStepFine.}

\item{abundStepFine}{Abundance step to use for the fine score array Must be smaller than abundStepCoarse.}

\item{SNR}{Signal-to-noise cutoff to use for peak picking. See ?MSnbase::pickPeaks.}

\item{method}{Method to use for peak picking. See ?MSnbase::pickPeaks.}

\item{refineMz}{Method for m/z refinement for peak picking. See ?MSnbase::pickPeaks.}

\item{k}{Number of neighboring signals to use for m/z refinement if refineMz = "kNeighbors". See ?MSnbase::pickPeaks.}

\item{resolvingPower}{Resolving power to be usedy for generating the initial theoretical spectrum. This parameter does not need to match the resolving power of the experimental spectrum.}

\item{compFunc}{Function to use for comparison of experimental and theoretical spectra. Acceptable values are "dotproduct" and "scoremfa".}

\item{binSize}{Bin size to use for peak binning prior to comparing spectra. See ?MSnbase::bin.}
}
\value{

}
\description{
calculate_score_array_DEV
}
