% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score_matrix.R
\name{calculate_score_matrix}
\alias{calculate_score_matrix}
\title{calculate_score_matrix}
\usage{
calculate_score_matrix(
  mz,
  intensity,
  sequence = NULL,
  charge = 1,
  start_12C = 0.98,
  start_14N = 0.98,
  abund_step = 0.001
)
}
\arguments{
\item{abund_step}{}
}
\value{

}
\description{
calculate_score_matrix
}
