% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate population size package.r
\name{CI}
\alias{CI}
\title{Calculate the confidence intervals of population size}
\usage{
CI(pop, detection, SDM, Match, plotrange = 10)
}
\arguments{
\item{pop}{A value of estimated population size.}

\item{detection}{1 / the average detection rate of distance sampling}

\item{SDM}{1 / the explained variance by the species distribution model}

\item{Match}{The standard deviation of the observation/prediction ratios at all panels of the whole area.}

\item{plotrange}{The plotting range for estimated population size.}
}
\value{
The 95\% confidence interval of the estimated population size
}
\description{
Evaluation the confidence intervals of the population estimation taking into account the uncertainty
of distance sampling, species distribution modelling, and prediction-observation matchness
}
\examples{

 CI(10000, 0.2, 0.8, 0.8)
 SDM = 1/max(RF$rsq)
 detection = 1/Average.p
 CI(EST[4], detection, SDM, sd.Match, 100)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
