% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate population size package.r
\name{cropLayers}
\alias{cropLayers}
\title{Crop raster layers to fit data in a smaller region}
\usage{
cropLayers(species, buffer, Envlayers)
}
\arguments{
\item{species}{A data.frame of species occurrences with columns "Name_E", "Lon", "Lat", "Date", "Count"}

\item{buffer}{A value of distance (unit: degree) defining the width of buffer zone outside the occurrences}

\item{Envlayers}{A RasterBrick (a multi-layer raster object)}
}
\value{

}
\description{
Crop raster layers to fit data in a smaller region using user defined buffer
}
\examples{
 attach(species) # load occurrences data
 BioClim <- brick('D:/GIS/CLIMATE/var27.grd')
 BioClim = cropLayers(species, buffer=0.5, Envlayers=BioClim)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
