% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot_outliers.R
\name{crossplot_outliers}
\alias{crossplot_outliers}
\title{To get the outlier data form a ggplot}
\usage{
crossplot_outliers(
  x,
  vars = NULL,
  num.outliers = NULL,
  percentile.outliers = NULL
)
}
\arguments{
\item{x}{a \code{ggplot2} plot or a \code{data.frame}. Can be
generated from \code{crossplot}.}

\item{vars}{the variables for which the outliers will be
calculated. If \code{x} is a \code{ggplot2} object, use the
variable names in the underlying data}

\item{num.outliers}{the number of outliers for each variable in
\code{vars}. Defaults to \code{NULL}}

\item{percentile.outliers}{the percentile used to measure outliers
for each variable in \code{vars}. Defaults to \code{NULL}}
}
\value{
a \code{data.frame} with the plot data for the outliers
}
\description{
This function is useful for getting outliers and labeling outlying
points on a plot. Based a \code{data.frame} or \code{ggplot} and
variables from the user, \code{crossplot_outliers} returns the
corresponding outliers. The user can input either the number of
desired outliers, \code{num.outliers}, or the percentile
corresponding to the outliers \code{percentile.outliers}, but not
both. Note: \code{NA} values will be replaced with the mean value.
}
\examples{
data(mtcars)
##Get the outliers for hp and and mpg from mtcars
crossplot_outliers(mtcars,vars = c("hp","mpg"), percentile.outliers = 0.05)
crossplot_outliers(mtcars,vars = c("hp","mpg"), num.outliers = 2)

##Outliers from a ggplot
p <- ggplot(mtcars, aes(x = hp, y = mpg)) + geom_point(aes(size = wt))
crossplot_outliers(p, vars = c("hp", "mpg"), num.outliers = 2)
crossplot_outliers(p, vars = c("mpg", "hp", "wt"), percentile.outliers = .05)

## -- Full Example using crossplot() and labels -- ##
mtcars$name <- rownames(mtcars)
mtcars.outliers <- crossplot_outliers(mtcars, vars = c("mpg", "hp"), num.outliers = 2)
p <- crossplot(mtcars, x.var = "mpg", y.var = "hp", size.var = "wt",
              shapes.var = "cyl", label.var = "name", points.alpha = 0.7) +
     geom_text(data = mtcars.outliers)
print(p)

##If there are missing values in vars, these will be replaced by the
##mean so they are less likely to affect the outliers
data(mtcars)
mtcars$mpg[7:12] <- rep(NA, 6)
##Get the outliers for hp and and mpg from mtcars
crossplot_outliers(mtcars,vars = c("hp","mpg"), percentile.outliers = 0.05)
crossplot_outliers(mtcars,vars = c("hp","mpg"), num.outliers = 2)
}
