% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate population size package.r
\name{estPopSize}
\alias{estPopSize}
\title{Estimate population size at the whole area}
\usage{
estPopSize(pop.pred, tracks, species, detection)
}
\arguments{
\item{pop.pred}{A raster layer with number of individuals at each cell}

\item{tracks}{A dataframe containing the Lat/Lon of thinned track points.}

\item{species}{A data.frame of species occurrences with columns "Name_E", "Lon", "Lat", "Date", "Count"}

\item{detection}{The average detection during distance sampling}
}
\value{
A vector including the predicted population size by Random Forest, predicted population on the routes,
Observed population on the routes, and final adjusted estimation of population size
}
\description{
Predict the population size using a species distribution model, and adjust the predicted value
based on real observations by distance sampling
}
\examples{

 EST = estPopSize(pred, tracks, species, Average.p); EST

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
