% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_mass_spectrum.R
\name{extract_mass_spectrum}
\alias{extract_mass_spectrum}
\title{extract_mass_spectrum}
\usage{
extract_mass_spectrum(
  rawFileDir = NULL,
  rawFileName = NULL,
  scanNum = NULL,
  mzRange = NULL
)
}
\arguments{
\item{mzRange}{}
}
\value{

}
\description{
extract_mass_spectrum
}
