% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_score_heatmap.R
\name{generate_score_heatmap}
\alias{generate_score_heatmap}
\title{generate_score_heatmap}
\usage{
generate_score_heatmap(
  score_matrix,
  fillOption = "C",
  scaleFillValues = c(0, 1),
  scaleFillLimits = c(0, signif(max(score_matrix), digits = 1)),
  compFunc = "dotproduct"
)
}
\arguments{
\item{score_matrix}{A matrix containing 12C and 14N abundances.}

\item{fillOption}{Argument passed to ggplot2::scale_fill_viridis_c. Controls which viridis color scale is used.}

\item{scaleFillValues}{Argument passed to ggplot2::scale_fill_viridis_c. Controls positions of colors along the gradient.}

\item{scaleFillLimits}{Argument passed to ggplot2::scale_fill_viridis_c. Controls values which are represented by the fill range.}

\item{compFunc}{Function used for comparison of experimental and theoretical spectra. This only impacts the legend label. Acceptable values are "dotproduct" and "scoremfa".}
}
\value{

}
\description{
generate_score_heatmap
}
