% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_array_max_names}
\alias{get_array_max_names}
\title{get_array_max_names}
\usage{
get_array_max_names(array)
}
\arguments{
\item{array}{}
}
\value{

}
\description{
get_array_max_names
}
\examples{

}
