% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_abundances.R
\name{get_optimal_abundances}
\alias{get_optimal_abundances}
\title{get_optimal_abundances}
\usage{
get_optimal_abundances(score_matrix)
}
\arguments{
\item{score_matrix}{A matrix containing 12C and 14N abundances.}
}
\value{

}
\description{
get_optimal_abundances
}
