% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot_outliers.R
\name{get_outlier_logicals}
\alias{get_outlier_logicals}
\title{To get the index of the outliers for a vector x}
\usage{
get_outlier_logicals(x, num.outliers = NULL, percentile.outliers = NULL)
}
\arguments{
\item{x}{a numeric vector}

\item{num.outliers}{the number of outliers. Defaults to \code{NULL}}

\item{percentile.outliers}{the percentile used to measure
outliers. Defaults to \code{NULL}}
}
\value{
a logical index where outliers are set to \code{TRUE}
}
\description{
This function will return a logical index where outliers are set to
\code{TRUE}. The user can imput either imput the number of desired
outliers, \code{num.outliers}, or the percentile corresponding to
the outliers \code{percentile.outliers}, but not both.
}
\details{
Note: \code{NA} values will be replaced with the mean value.
}
