% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spectrum_combined.R
\name{make_spectrum_combined}
\alias{make_spectrum_combined}
\title{make_spectrum_combined}
\usage{
make_spectrum_combined(
  mz = NULL,
  intensity = NULL,
  SNR = 10,
  method = "MAD",
  refineMz = "kNeighbors",
  k = 2,
  score_matrix = NULL,
  sequence = NULL,
  charge = NULL,
  isoAbundCutoff = 1,
  hClust_height = 0.005
)
}
\arguments{
\item{score_matrix}{}
}
\value{

}
\description{
make_spectrum_combined
}
