% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spectrum_peakpick.R
\name{make_spectrum_peakpick}
\alias{make_spectrum_peakpick}
\title{make_spectrum_peakpick}
\usage{
make_spectrum_peakpick(
  MSspectrum = NULL,
  mz = NULL,
  intensity = NULL,
  SNR = 10,
  method = "MAD",
  refineMz = "kNeighbors",
  k = 2,
  binSize = 0.05
)
}
\arguments{
\item{MSspectrum}{An MSnbase Spectrum1 object.}

\item{mz}{A numeric vector containing m/z values for a spectrum.}

\item{intensity}{A numeric vector containing intensity values for a spectrum.}

\item{SNR}{Signal-to-noise cutoff to use for peak picking. See ?MSnbase::pickPeaks.}

\item{method}{Method to use for peak picking. See ?MSnbase::pickPeaks.}

\item{refineMz}{Method for m/z refinement for peak picking. See ?MSnbase::pickPeaks.}

\item{k}{Number of neighboring signals to use for m/z refinement if refineMz = "kNeighbors". See ?MSnbase::pickPeaks.}

\item{binSize}{Bin size to use for peak binning prior to comparing spectra. See ?MSnbase::bin.}
}
\value{

}
\description{
make_spectrum_peakpick
}
