% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spectrum_theoretical.R
\name{make_spectrum_theoretical}
\alias{make_spectrum_theoretical}
\title{make_spectrum_theoretical}
\usage{
make_spectrum_theoretical(
  score_matrix,
  sequence = NULL,
  charge = NULL,
  isoAbundCutoff = 1,
  hClust_height = 0.005,
  xrange = NULL
)
}
\arguments{
\item{score_matrix}{}
}
\value{

}
\description{
make_spectrum_theoretical
}
