% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate population size package.r
\name{spatialMatch}
\alias{spatialMatch}
\title{Comparing prediction-observation ratio at different zones of the whole area}
\usage{
spatialMatch(species, buffer, tracks, pop.pred, grid = 4)
}
\arguments{
\item{species}{A data.frame of species occurrences with columns "Name_E", "Lon", "Lat", "Date", "Count"}

\item{buffer}{A value of distance (unit: degree) defining the width of buffer zone outside the occurrences}

\item{tracks}{A dataframe containing the Lat/Lon of thinned track points.}

\item{pop.pred}{A raster layer with number of individuals at each cell}

\item{grid}{A value defining the number of grids (grid^2) for prediction-observation comparison}
}
\value{
The standard deviation of the observation/prediction ratios at all panels of the whole area
}
\description{
Calculating the uncertainty of prediction-observation rate
}
\examples{

 sd.Match = spatialMatch(species, buffer, tracks, pop.pred, 6); sd.Match

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
