% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate population size package.r
\name{trackPoints}
\alias{trackPoints}
\title{Derive representative points from survey routes}
\usage{
trackPoints(shape)
}
\arguments{
\item{shape}{A GIS shapefile containing survey routes}
}
\value{
A dataframe containing the Lat/Lon of thinned track points.
}
\description{
This function derives track points, and thinning the points to the density of one point per km.
}
\examples{

 tracks = trackPoints(shape)

}
\author{
Xinhai Li (Xinhai_li_edu@126.com)
}
