% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score_matrix.R
\name{calculate_score_matrix}
\alias{calculate_score_matrix}
\title{calculate_score_matrix}
\usage{
calculate_score_matrix(
  MSspectrum = NULL,
  mz = NULL,
  intensity = NULL,
  sequence = NULL,
  charge = 1,
  start12C = 0.989,
  start14N = 0.995,
  abundStep = 0.001,
  SNR = 10,
  method = "MAD",
  refineMz = "kNeighbors",
  k = 2,
  resolvingPower = 3e+05,
  compFunc = "dotproduct",
  binSize = 0.05
)
}
\arguments{
\item{binSize}{}
}
\value{

}
\description{
calculate_score_matrix
}
