% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_score_matrix.R
\name{calculate_score_matrix}
\alias{calculate_score_matrix}
\title{calculate_score_matrix}
\usage{
calculate_score_matrix(
  mz,
  intensity,
  sequence = NULL,
  charge = 1,
  start12C = 0.98,
  start14N = 0.98,
  abundStep = 0.001,
  isoAbundCutoff = 5,
  binSize = 0.05
)
}
\arguments{
\item{binSize}{}
}
\value{

}
\description{
calculate_score_matrix
}
