% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_score_heatmap.R
\name{generate_score_heatmap}
\alias{generate_score_heatmap}
\title{generate_score_heatmap}
\usage{
generate_score_heatmap(score_matrix, fillOption = "C")
}
\arguments{
\item{score_matrix}{}
}
\value{

}
\description{
generate_score_heatmap
}
