% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_score_heatmap.R
\name{generate_score_heatmap}
\alias{generate_score_heatmap}
\title{generate_score_heatmap}
\usage{
generate_score_heatmap(
  score_matrix,
  fillOption = "C",
  scaleFillValues = c(0, 1),
  scaleFillLimits = c(0, signif(max(score_matrix), digits = 1)),
  compFunc = "dotproduct"
)
}
\arguments{
\item{compFunc}{}
}
\value{

}
\description{
generate_score_heatmap
}
