% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spectrum_peakpick.R
\name{make_spectrum_peakpick}
\alias{make_spectrum_peakpick}
\title{make_spectrum_peakpick}
\usage{
make_spectrum_peakpick(
  MSspectrum = NULL,
  mz = NULL,
  intensity = NULL,
  SNR = 10,
  method = "MAD",
  refineMz = "kNeighbors",
  k = 2,
  binSize = 0.05
)
}
\arguments{
\item{binSize}{}
}
\value{

}
\description{
make_spectrum_peakpick
}
